requirements.txt
pandas~=1.2.3
Flask~=1.1.2
numpy~=1.20.2
PyYAML~=5.4.1